Minor code clean up
