+++ auth/token_generator_test.go
@@ -252,11 +252,15 @@ func TestIAMSignerNoMetadataService(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	if _, err = signer.Email(ctx); err == nil {
-		t.Errorf("Email() = nil; want = error")
+	want := "failed to determine service account: "
+	_, err = signer.Email(ctx)
+	if err == nil || !strings.HasPrefix(err.Error(), want) {
+		t.Errorf("Email() = %v; want = %q", err, want)
 	}
-	if _, err = signer.Sign(ctx, []byte("input")); err == nil {
-		t.Errorf("Sign() = nil; want = error")
+
+	_, err = signer.Sign(ctx, []byte("input"))
+	if err == nil || !strings.HasPrefix(err.Error(), want) {
+		t.Errorf("Sign() = %v; want = %q", err, want)
 	}
 }
